+++ hal/i2c/hal_i2c.h
@@ -61,7 +61,7 @@ extern "C"
     struct hal_i2c_pin_cfg
     {
         uintptr_t port; /**< GPIOx_BASE */
-        uint32_t pin; /**< GPIO_PIN_x */
+        uint16_t pin; /**< GPIO_PIN_x */
         uint32_t clk_bus; /**< LL_AHBx_GRPy_PERIPH_GPIOx */
         uint32_t af; /**< GPIO_AFx_I2Cy */
         uint32_t output_type; /**< LL_GPIO_OUTPUT_* */
